requirements.txt
Flask==3.0.3
Jinja2==3.1.4
numpy==2.0.2
opencv-python==4.11.0.86
pillow==11.0.0
gunicorn==23.0.0
python-dotenv==1.0.1